requirements.txt
openpyxl>=3.1.2
ragflow-sdk>=0.8.0
requests>=2.31.0
python-dotenv>=1.0.0
PyPDF2>=3.0.1
pillow>=10.0.0
schedule>=1.2.0
